rafactor manipulation
