+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_namespace_functions.py
@@ -14,7 +14,6 @@ from ivy_tests.test_ivy.test_functional.test_core.test_statistical import (
 from ivy_tests.test_ivy.test_functional.test_core.test_linalg import (
     _get_dtype_value1_value2_axis_for_tensordot,
 )
-import ivy_tests.test_ivy.test_frontends.test_numpy.helpers as np_frontend_helpers
 
 
 # absolute
@@ -420,199 +419,6 @@ def test_jax_numpy_broadcast_to(
         on_device=on_device,
         arr=x[0],
         shape=shape,
-    )
-
-
-@st.composite
-def _get_clip_inputs(draw):
-    shape = draw(
-        helpers.get_shape(
-            min_num_dims=1, max_num_dims=5, min_dim_size=2, max_dim_size=10
-        )
-    )
-    x_dtype, x = draw(
-        helpers.dtype_and_values(
-            available_dtypes=helpers.get_dtypes("numeric"),
-            shape=shape,
-            min_value=-1e10,
-            max_value=1e10,
-        )
-    )
-    min = draw(st.booleans())
-    if min:
-        max = draw(st.booleans())
-        min = draw(
-            helpers.array_values(
-                dtype=x_dtype[0], shape=shape, min_value=-50, max_value=5
-            )
-        )
-        max = (
-            draw(
-                helpers.array_values(
-                    dtype=x_dtype[0], shape=shape, min_value=6, max_value=50
-                )
-            )
-            if max
-            else None
-        )
-    else:
-        min = None
-        max = draw(
-            helpers.array_values(
-                dtype=x_dtype[0], shape=shape, min_value=6, max_value=50
-            )
-        )
-    return x_dtype, x, min, max
-
-
-# clip
-@handle_frontend_test(
-    fn_tree="jax.numpy.clip",
-    input_and_ranges=_get_clip_inputs(),
-)
-def test_jax_numpy_clip(
-    *,
-    input_and_ranges,
-    num_positional_args,
-    as_variable,
-    native_array,
-    with_out,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    x_dtype, x, min, max = input_and_ranges
-    helpers.test_frontend_function(
-        input_dtypes=x_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        a_min=min,
-        a_max=max,
-    )
-
-
-# reshape
-@st.composite
-def _get_input_and_reshape(draw):
-    shape = draw(
-        helpers.get_shape(
-            min_num_dims=2, max_num_dims=5, min_dim_size=2, max_dim_size=10
-        )
-    )
-    x_dtype, x = draw(
-        helpers.dtype_and_values(
-            available_dtypes=helpers.get_dtypes("valid"),
-            min_num_dims=1,
-            max_num_dims=5,
-            min_dim_size=2,
-            max_dim_size=10,
-            shape=shape,
-        )
-    )
-    new_shape = shape[1:] + (shape[0],)
-    return x_dtype, x, new_shape
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.reshape",
-    input_x_shape=_get_input_and_reshape(),
-    order=st.sampled_from(["C", "F"]),
-)
-def test_jax_numpy_reshape(
-    *,
-    input_x_shape,
-    order,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    x_dtype, x, shape = input_x_shape
-    helpers.test_frontend_function(
-        input_dtypes=x_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        newshape=shape,
-        order=order,
-    )
-
-
-@st.composite
-def _arrays_idx_n_dtypes(draw):
-    num_dims = draw(st.shared(helpers.ints(min_value=1, max_value=4), key="num_dims"))
-    num_arrays = draw(
-        st.shared(helpers.ints(min_value=2, max_value=4), key="num_arrays")
-    )
-    common_shape = draw(
-        helpers.lists(
-            arg=helpers.ints(min_value=2, max_value=3),
-            min_size=num_dims - 1,
-            max_size=num_dims - 1,
-        )
-    )
-    unique_idx = draw(helpers.ints(min_value=0, max_value=num_dims - 1))
-    unique_dims = draw(
-        helpers.lists(
-            arg=helpers.ints(min_value=2, max_value=3),
-            min_size=num_arrays,
-            max_size=num_arrays,
-        )
-    )
-    xs = list()
-    input_dtypes = draw(
-        helpers.array_dtypes(available_dtypes=draw(helpers.get_dtypes("valid")))
-    )
-    for ud, dt in zip(unique_dims, input_dtypes):
-        x = draw(
-            helpers.array_values(
-                shape=common_shape[:unique_idx] + [ud] + common_shape[unique_idx:],
-                dtype=dt,
-            )
-        )
-        xs.append(x)
-    return xs, input_dtypes, unique_idx
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.concatenate",
-    xs_n_input_dtypes_n_unique_idx=_arrays_idx_n_dtypes(),
-)
-def test_jax_numpy_concat(
-    *,
-    xs_n_input_dtypes_n_unique_idx,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    xs, input_dtypes, unique_idx = xs_n_input_dtypes_n_unique_idx
-    helpers.test_frontend_function(
-        input_dtypes=input_dtypes,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        arrays=xs,
-        axis=unique_idx,
     )
 
 
@@ -1565,119 +1371,6 @@ def test_jax_numpy_bitwise_xor(
     )
 
 
-# moveaxis
-@handle_frontend_test(
-    fn_tree="jax.numpy.moveaxis",
-    dtype_and_a=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_value=-100,
-        max_value=100,
-        shape=st.shared(
-            helpers.get_shape(
-                min_num_dims=1,
-                max_num_dims=3,
-                min_dim_size=1,
-                max_dim_size=3,
-            ),
-            key="a_s_d",
-        ),
-    ),
-    source=helpers.get_axis(
-        allow_none=False,
-        unique=True,
-        shape=st.shared(
-            helpers.get_shape(
-                min_num_dims=1,
-                max_num_dims=3,
-                min_dim_size=1,
-                max_dim_size=3,
-            ),
-            key="a_s_d",
-        ),
-        min_size=1,
-        force_int=True,
-    ),
-    destination=helpers.get_axis(
-        allow_none=False,
-        unique=True,
-        shape=st.shared(
-            helpers.get_shape(
-                min_num_dims=1,
-                max_num_dims=3,
-                min_dim_size=1,
-                max_dim_size=3,
-            ),
-            key="a_s_d",
-        ),
-        min_size=1,
-        force_int=True,
-    ),
-)
-def test_jax_numpy_moveaxis(
-    *,
-    dtype_and_a,
-    source,
-    destination,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, a = dtype_and_a
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=a[0],
-        source=source,
-        destination=destination,
-    )
-
-
-# flipud
-@handle_frontend_test(
-    fn_tree="jax.numpy.flipud",
-    dtype_and_m=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_value=-100,
-        max_value=100,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-    ),
-)
-def test_jax_numpy_flipud(
-    *,
-    dtype_and_m,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, m = dtype_and_m
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        m=m[0],
-    )
-
-
 # power
 @handle_frontend_test(
     fn_tree="jax.numpy.power",
@@ -2544,41 +2237,6 @@ def test_jax_numpy_any(
     )
 
 
-# transpose
-@handle_frontend_test(
-    fn_tree="jax.numpy.transpose",
-    array_and_axes=np_frontend_helpers._array_and_axes_permute_helper(
-        min_num_dims=0,
-        max_num_dims=5,
-        min_dim_size=0,
-        max_dim_size=10,
-    ),
-)
-def test_jax_numpy_transpose(
-    *,
-    array_and_axes,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    array, dtype, axes = array_and_axes
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=array,
-        axes=axes,
-    )
-
-
 # diag
 @st.composite
 def _diag_helper(draw):
@@ -2631,46 +2289,6 @@ def test_jax_numpy_diag(
     )
 
 
-# flip
-@handle_frontend_test(
-    fn_tree="jax.numpy.flip",
-    dtype_value=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid", full=True),
-        shape=st.shared(helpers.get_shape(min_num_dims=1), key="value_shape"),
-    ),
-    axis=helpers.get_axis(
-        shape=st.shared(helpers.get_shape(min_num_dims=1), key="value_shape"),
-        min_size=1,
-        max_size=1,
-        force_int=True,
-    ),
-)
-def test_jax_numpy_flip(
-    *,
-    dtype_value,
-    axis,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    dtype, value = dtype_value
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        m=value[0],
-        axis=axis,
-    )
-
-
 # sqrt
 @handle_frontend_test(
     fn_tree="jax.numpy.sqrt",
@@ -2697,38 +2315,6 @@ def test_jax_numpy_sqrt(
         fn_tree=fn_tree,
         on_device=on_device,
         x=x[0],
-    )
-
-
-# fliplr
-@handle_frontend_test(
-    fn_tree="jax.numpy.fliplr",
-    dtype_and_m=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_num_dims=2,
-    ),
-)
-def test_jax_numpy_fliplr(
-    *,
-    dtype_and_m,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, m = dtype_and_m
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        m=m[0],
     )
 
 
@@ -3338,44 +2924,6 @@ def test_jax_numpy_diagonal(
     )
 
 
-# expand_dims
-@handle_frontend_test(
-    fn_tree="jax.numpy.expand_dims",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        shape=st.shared(helpers.get_shape(), key="expand_dims_axis"),
-    ),
-    axis=helpers.get_axis(
-        shape=st.shared(helpers.get_shape(), key="expand_dims_axis"),
-    ),
-)
-def test_jax_expand_dims(
-    *,
-    dtype_and_x,
-    axis,
-    with_out,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        axis=axis,
-    )
-
-
 # degrees
 @handle_frontend_test(
     fn_tree="jax.numpy.degrees",
@@ -3480,46 +3028,6 @@ def test_jax_numpy_asarray(
     )
 
 
-# take
-@handle_frontend_test(
-    fn_tree="jax.numpy.take",
-    dtype_indices_axis=helpers.array_indices_axis(
-        array_dtypes=helpers.get_dtypes("numeric"),
-        indices_dtypes=helpers.get_dtypes("integer"),
-        min_num_dims=1,
-        max_num_dims=5,
-        min_dim_size=1,
-        max_dim_size=10,
-        indices_same_dims=True,
-    ),
-)
-def test_jax_numpy_take(
-    *,
-    dtype_indices_axis,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtypes, value, indices, axis, _ = dtype_indices_axis
-    helpers.test_frontend_function(
-        input_dtypes=input_dtypes,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=3,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=value,
-        indices=indices,
-        axis=axis,
-    )
-
-
 # zeros_like
 @handle_frontend_test(
     fn_tree="jax.numpy.zeros_like",
@@ -3559,46 +3067,6 @@ def test_numpy_zeros_like(
         a=x[0],
         dtype=dtype[0],
         shape=shape,
-    )
-
-
-# stack
-@handle_frontend_test(
-    fn_tree="jax.numpy.stack",
-    dtype_values_axis=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("float"),
-        num_arrays=st.shared(helpers.ints(min_value=2, max_value=4), key="num_arrays"),
-        shape=helpers.get_shape(min_num_dims=1),
-        shared_dtype=True,
-        valid_axis=True,
-        allow_neg_axes=True,
-        force_int_axis=True,
-    ),
-    dtype=helpers.get_dtypes("valid", full=False),
-)
-def test_jax_numpy_stack(
-    dtype_values_axis,
-    dtype,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, values, axis = dtype_values_axis
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        arrays=values,
-        axis=axis,
     )
